+++ src/Traits/Observerable.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace CustomD\LaravelHelpers\Traits;
+
+trait Observerable
+{
+
+    public static function bootObserverable(): void
+    {
+
+        $observers = self::$observers ?? [];
+
+        $class = str(class_basename(self::class))->append('Observer')->prepend('\\App\\Observers\\')->__toString();
+        if (class_exists($class)) {
+            array_unshift($observers, $class);
+        }
+
+        foreach (array_unique($observers) as $observer) {
+            self::observe($observer);
+        }
+    }
+}